chapter-1: separates controller for welcome and ingredients, adds save ingredient in controller, corrects endpoints
